+++ Drivers/sx1262/sx1262_diag.h
@@ -14,7 +14,7 @@ extern "C" {
 uint32_t bandwidth2num(BandWidth_t bandwidth);
 bool parse_dev_stat(uint8_t dev_stat);
 bool parse_irq_stat(uint16_t irq_stat);
-char* bit_rate2str(float bit_s);
+char* bit_rate2str(double bit_s);
 bool parse_op_error(uint16_t op_error);
 bool print_int_diag(Sx1262IrqCnt_t* irq_cnt);
 bool printf_pack_stat(PaketStat_t* pack_stat, char* name);
