+++ Drivers/sx1262/sx1262_drv.h
@@ -190,7 +190,7 @@ typedef struct xSx1262_t {
 #ifdef HAS_SX1262_BIT_RATE
     uint32_t tx_start_time_stamp_ms;
     uint32_t tx_last_size;
-    float tx_max_bit_rate;
+    double tx_max_bit_rate;
 #endif /*HAS_SX1262_BIT_RATE*/
     uint32_t tx_done_time_stamp_ms;
     uint32_t busy_cnt;
